+++ packages/components/src/index.js
@@ -48,3 +48,4 @@ export { default as Image } from './image';
 export { default as Markdown } from './markdown';
 export { default as Card } from './card';
 export { default as Summary } from './summary';
+export { default as Switch } from './switch';