+++ packages/components/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [28.2.0](https://github.com/transferwise/neptune-web/compare/@transferwise/components@28.1.1...@transferwise/components@28.2.0) (2020-07-23)
+
+
+### Features
+
+* switch component ([#502](https://github.com/transferwise/neptune-web/issues/502)) ([cf43f4e](https://github.com/transferwise/neptune-web/commit/cf43f4e998646e7fdb49cb9267df25ae3195689b))
+
+
+
+
+
+## [28.1.1](https://github.com/transferwise/neptune-web/compare/@transferwise/components@28.1.0...@transferwise/components@28.1.1) (2020-07-20)
+
+**Note:** Version bump only for package @transferwise/components
+
+
+
+
+
+# [28.1.0](https://github.com/transferwise/neptune-web/compare/@transferwise/components@28.0.0...@transferwise/components@28.1.0) (2020-07-17)
+
+
+### Features
+
+* expose complete step of upload component ([#490](https://github.com/transferwise/neptune-web/issues/490)) ([82991a0](https://github.com/transferwise/neptune-web/commit/82991a0b95904d62f0918567be0af96e49aa38b4))
+
+
+
+
+
+# [28.0.0](https://github.com/transferwise/neptune-web/compare/@transferwise/components@27.14.0...@transferwise/components@28.0.0) (2020-07-14)
+
+
+* feat!: definition list improvements and more (#476) ([c330ed9](https://github.com/transferwise/neptune-web/commit/c330ed93363feb0864f3636c9fc5e79399df6d20)), closes [#476](https://github.com/transferwise/neptune-web/issues/476)
+
+
+### BREAKING CHANGES
+
+* Breaking changes in definition list
+
+
+
+
+
 # [27.14.0](https://github.com/transferwise/neptune-web/compare/@transferwise/components@27.13.2...@transferwise/components@27.14.0) (2020-07-14)