+++ packages/less-config/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://github.com/transferwise/neptune-web/compare/@transferwise/less-config@1.0.4...@transferwise/less-config@2.0.0) (2020-07-14)
+
+
+* feat!: definition list improvements and more (#476) ([c330ed9](https://github.com/transferwise/neptune-web/commit/c330ed93363feb0864f3636c9fc5e79399df6d20)), closes [#476](https://github.com/transferwise/neptune-web/issues/476)
+
+
+### BREAKING CHANGES
+
+* Breaking changes in definition list
+
+
+
+
+
 ## [1.0.4](https://github.com/transferwise/neptune-web/compare/@transferwise/less-config@1.0.3...@transferwise/less-config@1.0.4) (2020-07-10)